refactor(web): rename clerk_bindings → js_bridge (#490)
The file was originally Clerk-only, but PR 1+2+3 (Sentry foundation)
added `sentry_set_user`, `sentry_clear_user`, and `sentry_breadcrumb`
helpers to the same `#[wasm_bindgen(inline_js)]` block — they share
one snippet file because wasm-bindgen emits one snippet per inline_js
attribute, and consolidating reduces churn for any tool that has to
content-hash snippet folders.

The name `clerk_bindings.rs` is now misleading: it suggests pure Clerk
auth, but the file is a general JS bridge (Clerk + Sentry + future
window globals).

Mechanical rename:
- `git mv crates/intrada-web/src/clerk_bindings.rs js_bridge.rs`
- `pub mod clerk_bindings` → `pub mod js_bridge` in lib.rs
- All imports + call sites swept (`clerk_bindings::` → `js_bridge::`)
- File-level docstring rewritten to reflect the broader scope
- JS-side function names (`js_init_clerk`, `js_get_token`, etc.) kept
  intact — they describe what the functions do, not the module they
  live in

Verified: `cargo fmt --all -- --check`, `cargo clippy --workspace
-- -D warnings`, `cargo clippy --target wasm32-unknown-unknown -p
intrada-web -- -D warnings`, `cargo test --workspace` (452 pass) all
clean.

Closes #490
