+++ crates/intrada-web/src/components/profile_button.rs
@@ -1,7 +1,7 @@
 use leptos::prelude::*;
 use leptos_router::components::A;
 
-use intrada_web::clerk_bindings;
+use intrada_web::js_bridge;
 
 /// Avatar button that links to the Settings route.
 ///
@@ -12,7 +12,7 @@ use intrada_web::clerk_bindings;
 #[component]
 pub fn ProfileButton() -> impl IntoView {
     let initial = move || {
-        clerk_bindings::email()
+        js_bridge::email()
             .as_deref()
             .and_then(|s| s.chars().next())
             .map(|c| c.to_ascii_uppercase().to_string())
@@ -20,7 +20,7 @@ pub fn ProfileButton() -> impl IntoView {
     };
 
     view! {
-        <Show when=move || clerk_bindings::is_signed_in()>
+        <Show when=move || js_bridge::is_signed_in()>
             <A
                 href="/settings"
                 attr:class="ml-2 flex items-center justify-center h-8 w-8 rounded-full bg-surface-primary border border-border-default text-sm font-medium text-primary hover:bg-surface-hover focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-accent-focus motion-safe:transition-colors no-underline"
