+++ crates/intrada-web/src/components/session_summary.rs
@@ -6,8 +6,8 @@ use intrada_web::validation::validate_achieved_tempo_input;
 use crate::components::{
     AccentBar, Button, ButtonVariant, Icon, IconName, RatingChips, SetSaveForm, StatCard, StatTone,
 };
-use intrada_web::clerk_bindings;
 use intrada_web::core_bridge::process_effects;
+use intrada_web::js_bridge;
 use intrada_web::types::{IsLoading, IsSubmitting, SharedCore};
 
 /// End-of-session summary component: shows results, allows notes, save/discard.
@@ -344,7 +344,7 @@ pub fn SessionSummary() -> impl IntoView {
                                         full_width=true
                                         on_click=Callback::new(move |_| {
                                             let now = chrono::Utc::now();
-                                            clerk_bindings::sentry_breadcrumb("session", "session.save", "info");
+                                            js_bridge::sentry_breadcrumb("session", "session.save", "info");
                                             let event = Event::Session(SessionEvent::SaveSession { now });
                                             let core_ref = core_save.borrow();
                                             let effects = core_ref.process_event(event);
@@ -357,7 +357,7 @@ pub fn SessionSummary() -> impl IntoView {
                                 <Button
                                     variant=ButtonVariant::DangerOutline
                                     on_click=Callback::new(move |_| {
-                                        clerk_bindings::sentry_breadcrumb("session", "session.discard", "info");
+                                        js_bridge::sentry_breadcrumb("session", "session.discard", "info");
                                         let event = Event::Session(SessionEvent::DiscardSession);
                                         let core_ref = core_discard.borrow();
                                         let effects = core_ref.process_event(event);
