+++ crates/intrada-web/src/components/session_timer.rs
@@ -10,8 +10,8 @@ use crate::components::{
     Button, ButtonSize, ButtonVariant, Icon, IconName, InlineTypeIndicator, ItemReflectionSheet,
     ItemReflectionTarget, ProgressRing, SectionLabel, SetlistEntryRow, TransitionPrompt,
 };
-use intrada_web::clerk_bindings;
 use intrada_web::core_bridge::process_effects;
+use intrada_web::js_bridge;
 use intrada_web::types::{IsLoading, IsSubmitting, ItemType, SharedCore};
 
 /// Map an `ItemKind` from core into the `ItemType` enum used by
@@ -502,7 +502,7 @@ pub fn SessionTimer() -> impl IntoView {
                                         a.current_position == a.total_items.saturating_sub(1)
                                     });
                                     let event = if is_last_now {
-                                        clerk_bindings::sentry_breadcrumb(
+                                        js_bridge::sentry_breadcrumb(
                                             "session",
                                             "session.finish",
                                             "info",
