+++ crates/intrada-web/src/views/welcome.rs
@@ -5,7 +5,7 @@ use leptos_router::NavigateOptions;
 
 use crate::app::AuthState;
 use crate::components::{Button, ButtonVariant};
-use intrada_web::clerk_bindings;
+use intrada_web::js_bridge;
 use intrada_web::platform::is_ios;
 
 /// Public marketing homepage at `/`.
@@ -104,7 +104,7 @@ fn WelcomeNav() -> impl IntoView {
     let on_sign_in = Callback::new(move |_| {
         signing_in.set(true);
         leptos::task::spawn_local(async move {
-            clerk_bindings::sign_in_with_google().await;
+            js_bridge::sign_in_with_google().await;
         });
     });
 
@@ -147,7 +147,7 @@ fn WelcomeHero() -> impl IntoView {
     let on_sign_in = Callback::new(move |_| {
         signing_in.set(true);
         leptos::task::spawn_local(async move {
-            clerk_bindings::sign_in_with_google().await;
+            js_bridge::sign_in_with_google().await;
         });
     });
 
@@ -328,7 +328,7 @@ fn WelcomeFinalCta() -> impl IntoView {
     let on_sign_in = Callback::new(move |_| {
         signing_in.set(true);
         leptos::task::spawn_local(async move {
-            clerk_bindings::sign_in_with_google().await;
+            js_bridge::sign_in_with_google().await;
         });
     });
